fix: TASK-12 correcting mistakes after PR comments.Deleted tabPanel code and used TabPanel from directory Elements
https://trello.com/c/LEs0aFKy/24-task-12-add-main-page
