+++ src/components/main-page/tab-panel/TabPanelDescription.jsx
@@ -1,19 +1,57 @@
-import Typography from "@material-ui/core/Typography";
-import Box from "@material-ui/core/Box";
-import React from "react";
+import React from 'react';
+import PropTypes from 'prop-types';
+import tabPanelStyles from "./TabPanelStyles";
+import AppBar from '@material-ui/core/AppBar';
+import Tabs from '@material-ui/core/Tabs';
+import Tab from '@material-ui/core/Tab';
+import data from "../../../data/data";
+import AuthorOfDay from "../author-of-day/AuthorOfDay";
+import TeamList from "../../team-page/TeamList";
+import TabPanel  from "../../elements/ScrollableTabs/TabPanel/TabPanel";
+
+function a11yProps(index) {
+    return {
+        id: `simple-tab-${index}`,
+        'aria-controls': `simple-tabpanel-${index}`,
+    };
+}
+
+TabPanelDescription.propTypes = {
+    children: PropTypes.node,
+    index: PropTypes.any,
+    value: PropTypes.any,
+    mainPageData: PropTypes.object
+};
+TabPanelDescription.defaultProps = {
+    mainPageData: data.mainPage,
+};
 
 export default function TabPanelDescription(props) {
-    const {children, value, index, ...other} = props;
+    const classes = tabPanelStyles();
+    const [value, setValue] = React.useState(0);
+    const  description = props.mainPageData.description;
+    const handleChange = (event, newValue) => {
+        setValue(newValue);
+    };
+
     return (
-        <Typography
-            component="div"
-            role="tabpanel"
-            hidden={value !== index}
-            id={`simple-tabpanel-${index}`}
-            aria-labelledby={`simple-tab-${index}`}
-            {...other}
-        >
-            {value === index && <Box p={3}>{children}</Box>}
-        </Typography>
+        <div className={classes.root} >
+            <AppBar position="static">
+                <Tabs value={value} onChange={handleChange} aria-label="simple tabs example">
+                    <Tab label="Intro" {...a11yProps(0)} />
+                    <Tab label="Author of the day" {...a11yProps(1)} />
+                    <Tab label="Developer team" {...a11yProps(2)} />
+                </Tabs>
+            </AppBar>
+            <TabPanel activeTabIndex={value} index={0}>
+                {description}
+            </TabPanel>
+            <TabPanel activeTabIndex={value} index={1}>
+                <AuthorOfDay/>
+            </TabPanel>
+            <TabPanel activeTabIndex={value} index={2}>
+                <TeamList/>
+            </TabPanel>
+        </div>
     );
 }
